refactor: register폴더 이름변경 및 hooks 폴더 생성 및 위치 변경
