+++ src/pages/register/hooks/usePasswordValidation.ts
