+++ src/pages/register/types.ts
