+++ src/stories/RegisterPage.stories.tsx
@@ -1,0 +1,34 @@
+import { Meta, StoryObj } from '@storybook/react';
+import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
+
+import LayoutProvider from '~/common/components/Layout';
+import register from '~/pages/register';
+
+const queryClient = new QueryClient({
+  defaultOptions: {
+    queries: {
+      staleTime: 1000 * 60 * 5,
+      gcTime: 1000 * 60 * 5,
+      retry: 0,
+      refetchOnWindowFocus: false,
+    },
+  },
+});
+
+export default {
+  title: 'Pages/RegisterPage',
+  component: register,
+  argTypes: {},
+  args: {},
+  decorators: Story => (
+    <QueryClientProvider client={queryClient}>
+      <LayoutProvider>
+        <Story />
+      </LayoutProvider>
+    </QueryClientProvider>
+  ),
+} satisfies Meta<typeof register>;
+
+type Story = StoryObj<typeof register>;
+
+export const Default: Story = {};
